+++ examples/03-dev/demo/lila.config.js
@@ -67,6 +67,7 @@ module.exports = {
     },
     import: [{ "libraryName": "antd", style: "css" }],
     enableCssModules: !0,
-    cssModulesExclude: [/node_modules/, /src\/less/]
+    cssModulesExclude: [/node_modules/, /src\/less/],
+    browsers: ['> 1%', 'iOS 7']
 };