+++ project_config/webpack/loaders/extract_less_loader.js
@@ -1,7 +1,8 @@
 
 'use strict';
 
-var ExtractTextPlugin = require('extract-text-webpack-plugin');
+const ExtractTextPlugin = require('extract-text-webpack-plugin');
+const autoprefixer = require('autoprefixer');
 
 /**
  * make extract less loader
@@ -10,9 +11,10 @@ var ExtractTextPlugin = require('extract-text-webpack-plugin');
  * @param excludeMatches
  * @param include
  * @param exclude
+ * @param browsers
  * @returns {{test: RegExp, use: *}}
  */
-module.exports = (useCssModules = !1, excludeMatches = [], include = !1, exclude = !1) => {
+module.exports = (useCssModules = !1, excludeMatches = [], include = !1, exclude = !1, browsers = []) => {
     let loader = {
         test: /\.less$/,
         use: ExtractTextPlugin.extract({
@@ -22,6 +24,14 @@ module.exports = (useCssModules = !1, excludeMatches = [], include = !1, exclude
                     loader: 'css-loader',
                     options: {
                         modules: useCssModules
+                    }
+                },
+                {
+                    loader: 'postcss-loader',
+                    options: {
+                        plugins: [
+                            autoprefixer({browsers})
+                        ]
                     }
                 },
                 {