+++ project_config/webpack/loaders/less_loader.js
@@ -1,3 +1,6 @@
+
+const autoprefixer = require('autoprefixer');
+
 /**
  * make less loader
  *
@@ -5,9 +8,10 @@
  * @param excludeMatches
  * @param include
  * @param exclude
+ * @param browsers
  * @returns {{test: RegExp, use: *[]}}
  */
-module.exports = (useCssModules = !1, excludeMatches = [], include = !1, exclude = !1) => {
+module.exports = (useCssModules = !1, excludeMatches = [], include = !1, exclude = !1, browsers = []) => {
     let loader = {
         test: /\.less$/,
         use: [
@@ -18,6 +22,14 @@ module.exports = (useCssModules = !1, excludeMatches = [], include = !1, exclude
                 loader: 'css-loader',
                 options: {
                     modules: useCssModules
+                }
+            },
+            {
+                loader: 'postcss-loader',
+                options: {
+                    plugins: [
+                        autoprefixer({browsers})
+                    ]
                 }
             },
             {