+++ project_config/webpack/util/make_module.js
@@ -21,47 +21,48 @@ module.exports = (config) => {
     var rules = [babelLoader, htmlLoader];
 
     let excludeMatches = config.cssModulesExclude;
+    let browsers = config.browsers;
     if (config.packCssSeparately) {
         if (config.enableCssModules && excludeMatches) {
             rules.push(
-                makeExtractCssLoader(!1, excludeMatches, !0, !1),
-                makeExtractCssLoader(!0, excludeMatches, !1, !0),
-                makeExtractLessLoader(!1, excludeMatches, !0, !1),
-                makeExtractLessLoader(!0, excludeMatches, !1, !0)
+                makeExtractCssLoader(!1, excludeMatches, !0, !1, browsers),
+                makeExtractCssLoader(!0, excludeMatches, !1, !0, browsers),
+                makeExtractLessLoader(!1, excludeMatches, !0, !1, browsers),
+                makeExtractLessLoader(!0, excludeMatches, !1, !0, browsers)
             );
         }
         else if (config.enableCssModules) {
             rules.push(
-                makeExtractCssLoader(!0, excludeMatches, !1, !1),
-                makeExtractLessLoader(!0, excludeMatches, !1, !1)
+                makeExtractCssLoader(!0, excludeMatches, !1, !1, browsers),
+                makeExtractLessLoader(!0, excludeMatches, !1, !1, browsers)
             );
         }
         else {
             rules.push(
-                makeExtractCssLoader(),
-                makeExtractLessLoader()
+                makeExtractCssLoader(!1, [], !1, !1, browsers),
+                makeExtractLessLoader(!1, [], !1, !1, browsers)
             );
         }
     }
     else {
         if (config.enableCssModules && excludeMatches) {
             rules.push(
-                makeCssLoader(!1, excludeMatches, !0, !1),
-                makeCssLoader(!0, excludeMatches, !1, !0),
-                makeLessLoader(!1, excludeMatches, !0, !1),
-                makeLessLoader(!0, excludeMatches, !1, !0)
+                makeCssLoader(!1, excludeMatches, !0, !1, browsers),
+                makeCssLoader(!0, excludeMatches, !1, !0, browsers),
+                makeLessLoader(!1, excludeMatches, !0, !1, browsers),
+                makeLessLoader(!0, excludeMatches, !1, !0, browsers)
             );
         }
         else if (config.enableCssModules) {
             rules.push(
-                makeCssLoader(!0, excludeMatches, !1, !1),
-                makeLessLoader(!0, excludeMatches, !1, !1)
+                makeCssLoader(!0, excludeMatches, !1, !1, browsers),
+                makeLessLoader(!0, excludeMatches, !1, !1, browsers)
             );
         }
         else {
             rules.push(
-                makeCssLoader(),
-                makeLessLoader()
+                makeCssLoader(!1, [], !1, !1, browsers),
+                makeLessLoader(!1, [], !1, !1, browsers)
             );
         }
     }